Add new "matrix setup wait" configuration option and make debug console more robust
Add new EEPROM value g_matrix_setup_wait as well as a config console
option to update it.  Fix the robustness of the config console by making
it wait for the user to lift off the Enter key.  Rebuild all targets,
update the manual, and bump the minor version number for the new
feature.