Track layout filename so one can File > Save/Build.
The UI will now track the currently open layout and offer File >
Save functionality that does not ask for a filename. It saves to
the current layout file that was opened. It also offers a File >
Build Firware menu entry that will build the firmware right next
to the layout filename.

Existing functionality of asking for the filename during save and
build was preserved in the new menu entires File > Save Layout
As... and File > Build Firmware As...