Update README.md
I'm running Atmel Flip 3.4.7 on Windows. Following the instructions exactly the device is not flashed correctly. What appears to happen is that the hex file, loaded in step 2, is lost and no longer shown next to "HEX File:" on the GUI. This occurs after connecting to the USB device. 

Change proposed should work for all versions, in theory.