Fix macro buffering
As the autokey code refilled the buffer from a macro etc., it would lift
any modifiers that happened to be down.  This would be bad if there is
an ALT code that happens to straddle a buffer boundary.  This should fix
it up.  Of course, there is still a vulnerability.  With this change, if
there is a modifier held at the end of a macro etc., it would be still
be active if there is another one queued up right behind it.  However, I
consider that unlikely so this should be okay.