Rework USB endpoints
Issues #21 and #73.

Completely rework the USB endpoints and how they are selected.  Media
and Power have been combined, and NKRO has been split out into a
separate endpoint.  This preserves the boot-compatible endpoint to be
used at any time.  The configuration is now done through the keymapper
instead of the config console, where it is much less cumbersome.  Also
add a bunch of cleanup.

This is a proof of concept.  In general, the new system works, but a few
things are left to be done.  In particular, the config console is
broken.